Dont predict when dt is very small
When dt is small (less than 0.001 for example), you can skip the
prediction step and effectively run multiple sequential update steps
for the 'same' timestamp.

This has the benefit of being a bit more efficient as it allows you to
avoid the extra prediction step calculations including the update of F_
and calculation of Q_ above.