Initializing the Hj properly
This method of returning Hj early without initialising the values of Hj
can cause instability problems for the filter when px and py are both
0. This is caused by returning a declared but un-initialised matrix.
The matrix won't necessarily be full of 0s and won't necessarily have
the same values on every execution. There are two approaches that I've
seen work successfully. You can return a matrix initialised with 0s.
Eigen includes a fill method so you could call Hj.fill(0.0);.

Alternatively, rather than returning Hj early like this, you could set
c1 to a small value when it is below your threshold for this condition
and continue with the calculation of the Jacobian matrix.
71